+++ .ycm_extra_conf.py
@@ -87,6 +87,22 @@ get_python_inc(),
 '-isystem',
 'cpp/ycm/benchmarks/benchmark/include',
 '-std=c++17',
+'-I/usr/include/gtk-2.0', 
+'-I/usr/lib/x86_64-linux-gnu/gtk-2.0/include',
+'-I/usr/include/pango-1.0', 
+'-I/usr/include/atk-1.0',
+'-I/usr/include/gdk-pixbuf-2.0',
+'-I/usr/include/libmount', 
+'-I/usr/include/blkid',
+'-I/usr/include/pango-1.0',
+'-I/usr/include/fribidi',
+'-I/usr/include/cairo', '-I/usr/include/pixman-1',
+'-I/usr/include/harfbuzz',
+'-I/usr/include/glib-2.0',
+'-I/usr/lib/x86_64-linux-gnu/glib-2.0/include',
+'-I/usr/include/uuid',
+'-I/usr/include/freetype2',
+'-I/usr/include/libpng16'
 ]
 
 # Set this to the absolute path to the folder (NOT the file!) containing the