+++ ChangeLog.md
@@ -10,6 +10,7 @@
 - Add shell file to use this project to setup vim.
 - Add plugins "vim-flog" and "vim-fugitive" to show git log graph in vim.
 - Add plugin "vim-gitgutter" to show the change status for the editing file.
+- Add flags into ".ycm_extra_conf.py" to include the lib "gtk+-2.0".
 
 ### Changed
 - Update the template for git-commit command.